fix: should export all files by index and initialize by core
